kafka-project/kafka-layer/consumer/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
